Remove qualifier from scientificName and add lowest level taxon rank that was determined.
